+++ CHANGES.rst
@@ -5,8 +5,8 @@ Changelog
 2.9.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
-
+- Updated afternoon time range.
+  [daniele]
 
 2.9.0 (2026-06-09)
 ------------------